[cxx-interop] Allows clang importer to visit partially invalid anonymous unions
At times when an anonymous union has contents that is invalid it can
cause an assert in the clang importer because the field named
__Unnamed_union___Anonymous_field is not generated because the importer
attempts to avoid visiting invalid decls. This allows for visiting the
decl for a narrow case so that getter/setter generation surrounding the
union does not result in an assert.