[CSSimplify] Increase fix impact when passing closure to a non-function type parameter (#37388)
In overloaded context it's possible that there is an overload that
expects a closure but it can have other issues e.g. different number
of parameters, so in order to pick a better solution let's always
increase a score for overloads where closure is matched against a
non-function type parameter. Also a note which points out where
exactly anonymous parameters are used in the multi-statement
body is helpful to identify problematic and fix the issue.

Resolves: rdar://76058892