[concurrency] Provide missing header file. (#37004)
HAVE_PTHREAD_H is supplied in this config file.

The `__has_include` pattern is discouraged, but there is no other way of
conditioning on `pthread_np.h`, which is required for `pthread_main_np`
on OpenBSD. Therefore, use it, but call it out in a comment.

(cherry picked from commit eb945c81f28954f25ab95309339af8fbf0269c17)