[irgen] Add a new experimental transform called StackSizeLimitTransform.
Today the decision on whether or not to stack allocate an alloc_ref is decided
in a hard to maintain manner in IRGen. This is because the implementation is
fighting with the nature of IRGen being a visitor and how that hamstrings the
ability to ignore stack slots that do not overlap. It also does not limit the
size of alloc_stack that we can create per allocation frame.

In this patch, I introduce (behind a flag) a new IRGenPass
(StackSizeLimitTransform) that attacks this problem by moving the conditional
logic out of IRGen into a SIL pass. In the SIL pass, I ported the logic from
IRGen that is used to determine if a value is stack allocated. (This required a
little bit of refactoring).

This pass (enabled behind a flag) takes a slightly different approach: leave the
decision on whether or not to stack allocate such a thing to an IRGenSILPass
that runs before IRGen and factors out this logic. Then IRGen just assumes that
if an alloc_ref is marked as being on the stack that it /must be/ on the stack.

The pass takes the logic used in IRGen and expands it to also include
alloc_stack. The current algorithm in IRGen is based around the notion that each
SILFunction has a fixed budget of stack space that we will allow (today: 1024
bytes). It works by:

1. Gathering up all alloc_stack in the function and subtracting their size from
   the total stack size budget.

2. When IRGening the alloc_ref [stack], we perform some structural checks and
   then see if this alloc_ref's storage would push us out of stack size budget
   for the function we are visiting. If it would, we do not stack allocate the
   ref and instead heap allocate it.

StackSizeLimitTransform does a similar thing, but extends it to alloc_stack and
implements the check as a transform of the IR by converting alloc_stack to
alloc_box and alloc_ref [stack] -> alloc_ref.

The pass also takes advantage of its knowledge that alloc_ref [stack] means the
value /will/ be on the stack to expand a new Builtin for the stdlib:
Builtin.isOnStack(cls : AnyObject) -> Builtin.Int1. This builtin is only meant
to help the stdlib work around an issue where the stdlib needs to run different
code depending on if a value is /really/ on the stack or not.

NOTE: In this commit, we still always run the old algorithm, but in a
forthcoming commit I am going to add a switch that will just assume that
canBeOnStack means 'mustBeOnStack'.

I am going to extract out the code duplication into shared code before this
commit goes in.