[NFC] Cleanup uncovered async diagnostic inserts
All of the work to insert can be done much more easily with just the
subscript operator. The densemap subscript will automatically initialize
an element at the key if it hasn't already been created. This way we
don't need to insert the empty vector at the anchor before emplacing it,
we can jump straight to grabbing the vector and doing the insertion.