[Diagnostics] NFC: Add asserts to verify from/to types in contextual failure
We have a couple of reports that `ContextualFailure` crashes in
`trySequenceSubsequenceFixIts` even with the check for a valid
stdlib was added, which can only mean that from/to types are
somehow invalid.

Let's add a couple of asserts to verify their presence to aid us
in determining what expressions can cause the problem.