[concurrency] Avoid capacity in Deque on OpenBSD.
Deque creates a new ManagedBuffer when growing the underlying storage for
its elements. It sets the capacity in the header object, however it
initializes that field with ManagedBuffer.capacity.

ManagedBuffer.capacity is marked unavailable on OpenBSD because
ManagedBuffer unavoidably depends on malloc introspection which is
unavailable on this platform.

Obviously, we could mark the whole class as unavailable, but we don't
need to do anything so drastic since it appears though that Deque doesn't
unavoidably need this introspection. Therefore, on OpenBSD, just use
`minimumCapacity` when creating a new ManagedBuffer when growing and
moving elements.